socialnet_faas/functions/image_thumbnail/requirements.txt
Pillow==11.3.0
boto3==1.39.0
